requirements.txt
Flask
tensorflow
numpy
Pillow
adversarial-robustness-toolbox
torch
torchvision